turn python generator into a subclass of generator
This change is large, but mostly mechanical. In generator.cc the prefix
'generator::' is added to turn static functions into class methods. In
python.cc functions are moved into the newly introduced
python_generator. This sometimes requires the introduction of a new
isl_class argument.

The only non-trivial change is the separation of generate_python into
generator::generator() and python_generator::generate(). The former now
constructs the set of isl_classses which we want to generate bindings
for, whereas the latter prints for each class the corresponding bindings.

This change does not change current behavior. The generated isl.py files
are identical.

Signed-off-by: Armin Größlinger <armin.groesslinger@uni-passau.de>
Signed-off-by: Andreas Simbuerger <simbuerg@fim.uni-passau.de>
Signed-off-by: Tobias Grosser <tobias@grosser.es>