add C++ interface generator
The generator creates for each exported isl object a smart pointer class
of the following form:

      class Set;
      inline Set manage(__isl_take isl_set *Set);

      class Set {
        friend inline Set manage(__isl_take isl_set *Set);

        isl_set *Ptr = nullptr;

        inline explicit Set(__isl_take isl_set *Ptr);

      public:
        inline Set();
        inline Set(const Set &Obj);
        inline Set& operator=(Set Obj);
        inline ~Set();
        inline __isl_give isl_set *copy() const &;
        inline __isl_give isl_set *copy() && = delete;
        inline __isl_keep isl_set *get() const;
        inline __isl_give isl_set *release();
        inline bool isNull() const;
      };

The interface and functionality resembles C++11 std::unique_ptr. The
isl smart pointer keeps unique ownership of an object. The currently
held pointer can be observed by calling get(). When calling release(),
both the current pointer is returned and ownership is released. Names
and semantics of get() and release() match the corresponding methods in
std::unique_ptr. In addition, a copy() method is provided, which returns
a pointer to a copy of the managed object. These three methods allow the
isl C++ objects to be used together with the isl C functions. On top of
this an isNull method is provided to check if the object is currently
NULL.

Example:

     void foo(Set S) {

         isl_set_is_universe(S.get());

         isl_set_free(S.copy());

         isl_set_free(S.release());
     }

The r-value version of the copy() function is not needed as copying an
object that will be freed right after is inefficient. In such a
situation, the use of release() is more efficient as it avoids the
unnecessary object copy. By deleting the r-value version of copy we
ensure that the user does not accidentally use the copy() in situations
where it is preferable to call release().

An alternative choice to name get() and release() would have been keep()
and take(). The name 'take' is for example used in llvm::OwningPtr. For
consistency with std::unique_ptr, this alternative naming has not been
chosen.

isl C++ objects can be constructed either by copying existing isl
objects of the same type or by passing an isl C pointer to a global
constructor method isl::manage(isl_type *ptr). Constructing an
isl C++ object from an isl C pointer using an explicit class constructor
is not allowed. This interface has been chosen in correspondence with
std::make_unique. Besides making object construction and especially
pointer consumption explicit, global constructor methods also make
specifying the needed object type unnecessary. We also allow default
construction of empty isl objects, as default-constructability is
important when using isl objects in C++ containers and especially
when exploiting C++ move semantics. To complete the "Big Three" we also
define the copy assignment operator. The remaining two functions needed
to complete the "Big Five", move constructor and move assignment, are not
needed for an initial implementation and are left for a subsequent
commit.

Example:

  void foo (__isl_take *s1, __isl_take *s2) {

    isl::Set S1 = isl::manage(s1);

    auto S = isl::manage(s2);

    auto S2 = S1;

    isl::Set EmptySet;
  }

The currently generated library is header-only, which resembles the
already existing python bindings and is the best choice for a
lightweight interface library. In the future, implementations for
larger functions or debugging methods might require the addition
of implementations as part of a compiled (not-header-only) library.

The currently generated library is generated in a single unified header
file, which matches the generated python bindings. In the future,
this might evolve to a combination of per-class header files and a
unifying header file, as there is interest in both. However, for
an initial binding a unified header has been chosen.

No class hierarchy is constructed, as there is currently no obvious
benefit and a hierarchy can always be introduced later. However,
most conversions are probably better implemented through implicit
constructors from related isl objects. Such constructors are planned
to be added in the future.

All functions are declared 'inline' to not cause link errors when
included in multiple translation units.

The public interfaces have currently no defined behavior in case isl
ptrs are NULL, but instead assert when used on a NULL pointer. We
explicitly do not support isl's ability to handle NULL pointers
gracefully to ensure no software relies on this feature, which could
block the possible introduction of a more C++ style error handling
approach later on.

This change currently only enables C++ interface generation for sets
and maps, as only this part has been thoroughly tested. Other isl
datatypes will be exposed in future commits.

This change intentionally only introduces a minimalistic interface. The
interface established with this change is already sufficient to manage
isl pointers. Polly today uses an isl C++ interface that has about the
same functionality as this change and already the simplified memory
management has proven to facilitate software development a lot. The
infrastructure established with this change will serve as basis for a
future function call interface that makes isl C functions conveniently
accessible through the C++ interface.

File IO is performed through ostream, as this makes it easy to split
the bindings into multiple files later on. However, we still use C
style string formatting as this is more readable and closer to the
python binding generator. This approach is similar to the string
generation in Armin Groesslinger's SCALA bindings.

The design of the proposed isl C++ smart pointer interface is the result
of discussions with Michael Kruse, Jakob Leben, and Alexandre Isoard. It
was inspired by earlier bindings developed by Michael, Jakob, and
Alexandre, but also the C++ bindings developed by Andreas Simbuerger,
the SCALA bindings from Armin Groesslinger, as well as islpy developed
by Andreas Kloeckner. The actual code generator has been newly
developed, but relies on three enabling patches from Armin Groesslinger,
which establish an infrastructure also used in Armin's and Andreas'
bindings. As a result, this change also will make it easier to upstream
Armin's SCALA bindings and the similar structure of this new code
generator should make it easy to port changes from Andreas' C++
generator.

Signed-off-by: Tobias Grosser <tobias@grosser.es>