cpp: derive methods for isl-object-only functions
Derive methods for the most simple isl functions, the ones that only
work on isl objects.

Definition:

  Set Set::subtract(const Set &set2) {
    return manage(isl_set_subtract(copy(), set2.copy()));
  }

Usage:

  isl_ctx *ctx = isl_ctx_alloc();
  isl_set *s1 = isl_set_read_from_str(ctx, "{[i] : }");
  isl_set *s2 = isl_set_read_from_str(ctx, "{[i] : i > 0 }");

  isl::Set S1(s1);
  isl::Set S2(s2);
  auto S3 = S1.subtract(S2);

All functions are declared 'inline' to avoid conflicts when the same
function is embedded into multiple translation units that are later
linked together.

Functions that take (as first argument) and return the same isl object
type could have been implemented as in-place operations. This was not
done to remain consistent with the feeling of C-isl, the python
bindings, as well as the C++ bindings for functions where a different
type is returned.  The same choice has also been taken in Andreas
Simbuerger's C++ bindings. In the future, in-place operatins may be
introduced additionally.

This change does not generate code for methods that match C++ keywords,
use enums, callbacks, would need to be generated as 'static' functions,
or have return types and parameters that are none-isl objects (e.g.,
integers).

Performance optimizations that exploit move semantics have also not been
included to focus on core functionality.

Signed-off-by: Tobias Grosser <tobias@grosser.es>