cpp: generate bindings for all isl classes
The set of classes exported has been kept minimalistic to limite the
scope of the initial C++ generator. To allow larger pieces of code
(parts of Polly) to be ported to the new C++ bindings, we widen the set
of exported isl classes to all classed marked with __isl_export.

Signed-off-by: Tobias Grosser <tobias@grosser.es>