cpp: print Obj::operator bool(), if available.
Provide for each object a conversion operator to type bool that
evaluates to true in case the object is a valid object (Obj.Ptr !=
nullptr).

            %s::operator bool() const

TODO: Is this really desirable or are there for certain types better
uses for operator bool(), e.g. to only be true if 'isl::Val' != 0?